- Aggiunti permessi Gruppi di Appartenenza; - Aggiunti permessi Utenti;

+++ app/Http/Livewire/Users/Edit.php
@@ -3,11 +3,14 @@
 	namespace App\Http\Livewire\Users;
 
 	use App\Models\User;
+	use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 	use LivewireUI\Modal\ModalComponent;
 	use Spatie\Permission\Models\Role;
 
 	class Edit extends ModalComponent
 	{
+		use AuthorizesRequests;
+
 		public User $user;
 		public $name;
 		public $email;
@@ -16,6 +19,7 @@
 		public $roles;
 
 		public function mount($id) {
+			$this->authorize('user_update');
 			$this->user = User::find($id);
 			$this->roles = Role::all([
 				'id',
@@ -37,6 +41,7 @@
 		}
 
 		public function delete() {
+			$this->authorize('user_delete');
 			$this->user->delete();
 			$this->emit('user-deleted');
 			$this->closeModal();
